Final adjustments for elemnt change dark mode
